+++ lib/src/actions/actions.dart
@@ -8,6 +8,7 @@ import 'package:layrz_models/layrz_models.dart';
 part 'actions.freezed.dart';
 part 'actions.g.dart';
 
+part 'src/variants.dart';
 part 'src/action_type.dart';
 part 'src/action_subtype.dart';
 part 'src/action_property.dart';