+++ lib/src/commands/src/command_input.dart
@@ -48,4 +48,53 @@ abstract class CommandInput with _$CommandInput {
   }) = _CommandInput;
 
   factory CommandInput.fromJson(Map<String, dynamic> json) => _$CommandInputFromJson(json);
+
+  // coverage:ignore-start
+  /// [save] is a helper function to save the command input to the server.
+  Future<StandardResponse<DeviceCommand>> save({
+    /// [apiToken] is the API token to use for the request.
+    required String apiToken,
+
+    /// [uri] is the URI of the GraphQL endpoint.
+    required Uri uri,
+
+    /// [onResponse] is a callback that is called when the response is received.
+    ValueChanged<ApiStatus>? onResponse,
+  }) async {
+    final connector = LayrzConnector(apiToken: apiToken, uri: uri);
+    try {
+      final response = await connector.mutate(
+        GqlMutation(
+          name: id == null ? 'addDeviceCommand' : 'editDeviceCommand',
+          variables: [
+            GqlVariable(
+              name: 'data',
+              value: toJson(),
+              type: .input(of: 'DeviceCommandInput'),
+              isRequired: true,
+            ),
+          ],
+          fields: [
+            GqlField(name: 'status'),
+            GqlField(name: 'errors'),
+            GqlField(name: 'result', fragment: DeviceCommand.fragment),
+          ],
+        ),
+        _deviceCommandDecoder,
+      );
+
+      if (response.status != .ok) {
+        Log.warning('layrz_models/CommandInput/save(): Error saving command input: ${response.status}');
+        onResponse?.call(response.status);
+        return (response.status, response.errors, null);
+      }
+
+      return (response.status, null, response.result);
+    } catch (err, trace) {
+      Log.critical('layrz_models/CommandInput/save(): Error saving command input: $err\n$trace');
+      return (ApiStatus.internalError, null, null);
+    }
+  }
+
+  // coverage:ignore-end
 }